Use single component for role manipulation
